BPE: contracting single-token string is no-op
Without this fix, functions contract_pair and contract_and_count_pairs
fail to consider the case where a string has compressed down to a single
token. Applying the usual processing to such a string results in the
generation of an uninitialized empty string, which is wrong and leads to
crashes (Numba bug).